refactor entry and entry block downloading
